Fix bug in ZhuNote.py to avoid duplicate entries
Master pickle was a list of dictionaries (lod), and new entry was
added by lod.append. If modify an existing entry, it duplicates
this entry in the list. Fix is replacing lod by dod (dictionary of
dictionaries). The master pickle now is a dictionary and the key to
each entry is the entry title. This overwrites existing entry.

The unique ID in the master pickle is the title of the entry.